requirements.txt
Flask~=3.1.0
Flask-SQLAlchemy
Flask-Migrate~=4.0.7
boto3~=1.35.81
requests~=2.32.3
Pillow~=11.0.0
icrawler~=0.6.9
python-dotenv~=1.0.1
celery~=5.4.0
redis
psycopg2-binary==2.9.10
Flask-Login~=0.6.3
Werkzeug~=3.1.3
WTForms~=3.2.1
alembic~=1.14.0
SQLAlchemy~=2.0.36